#2 Вынос tokenRefresher в отдельный пакет
